+++ qiskit/ignis/verification/quantum_volume/__init__.py
@@ -14,9 +14,9 @@
 
 
 """
-Coherence module
+Quantum volume module
 """
 
-# Measurement correction functions
-from .circuits import t1_circuits, t2_circuits, t2star_circuits
-from .fitters import T1Fitter, T2Fitter, T2StarFitter
+# Quantum volume functions
+from .circuits import qv_circuits
+from .fitters import QVFitter