+++ test/tomography/test_process_tomography.py
@@ -1,9 +1,16 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright 2019, IBM.
+# This code is part of Qiskit.
 #
-# This source code is licensed under the Apache License, Version 2.0 found in
-# the LICENSE.txt file in the root directory of this source tree.
+# (C) Copyright IBM 2019.
+#
+# This code is licensed under the Apache License, Version 2.0. You may
+# obtain a copy of this license in the LICENSE.txt file in the root directory
+# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# Any modifications or derivative works of this code must retain this
+# copyright notice, and modified files need to carry a notice indicating
+# that they have been altered from the originals.
 
 # pylint: disable=missing-docstring
 
@@ -25,8 +32,8 @@ def run_circuit_and_tomography(circuit, qubits):
     job = qiskit.execute(qst, Aer.get_backend('qasm_simulator'),
                          shots=5000)
     tomo_fit = tomo.ProcessTomographyFitter(job.result(), qst)
-    choi_cvx = tomo_fit.fit(method='cvx')
-    choi_mle = tomo_fit.fit(method='lstsq')
+    choi_cvx = tomo_fit.fit(method='cvx').data
+    choi_mle = tomo_fit.fit(method='lstsq').data
     return (choi_cvx, choi_mle, choi_ideal)